% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acoustics}
\alias{acoustics}
\title{Acoustic measurements for the UCLA voice quality project}
\format{A data frame with 151,488 rows and 66 variables:
\describe{
  \item{speaker_id}{Speaker ID. Unique across all eight languages.}
  \item{speaker_number}{Speaker number or pseudonym. Unique within a language
  but not across languages.}
  \item{speaker_sex}{F for female or M for male.}
  \item{language}{Language that this word is from.}
  \item{language_variety}{Dialect of the language, or the location where this
  word was recorded.}
  \item{filename}{Filename of the corresponding audio file. For a description
  of file naming conventions, see the PDF notes for each language on the
  project website.}
  \item{segment_start}{Start time of this phonetic segment in the audio file,
  in milliseconds.}
  \item{segment_end}{End time of this phonetic segment in the audio file, in
  milliseconds.}
  \item{duration}{Duration of this segment, in milliseconds.}
  \item{timepoint}{Time within the segment where these measurements were
  taken, as a number from 1 to 9. A 1 means that this row has measurements
  taken during the first ninth of the segment, a 2 means that it has
  measurements taken during the second ninth, etc.}
  \item{textgrid_label}{Segment label in the corresponding Praat TextGrid.
  See the PDF notes for each language on the project website for IPA
  equivalents.}
  \item{segment_type}{Whether the segment is a vowel or a consonant.}
  \item{vowel_quality}{Vowel quality for this segment. NA if these
  measurements are from a consonant.}
  \item{phonation}{Phonation type of the segment.}
  \item{vowel_nasality}{Whether the vowel is oral or nasal.}
  \item{tone_contour}{Sequence of tone targets for this tone type, using the
  standard IPA range with the numbers 1-5, where 1 is low, 3 is mid, and 5 is
  high. NA for languages without lexical tone.}
  \item{number_of_tone_targets}{Number of tone targets for this tone type,
  from 1 to 3.}
  \item{tone_category}{Label for tone contour, divided into four categories:
  high (55, 53, 45, 44, 535), mid (24, 33, 35, 31), low (22, 21, 11, 213,
  13), and big (51, 351, 15, 151, 153).}
  \item{language_specific_tone}{Tone label based on language-specific
  tone transcription conventions, which were also used to annotate tone in
  the original TextGrid file. For example, in Mandarin Chinese, the
  files were annotated with the traditional labels of Tone 1, 2, 3 or 4.
  The \code{tone_contour} and \code{tone_category} columns have the actual
  tone contours for each tone type.}
  \item{consonant_aspiration}{Whether the consonant preceding the vowel is
  aspirated or unaspirated. If this segment is a consonant, then whether this
  segment is aspirated or unaspirated.}
  \item{consonant_nasality}{Whether the consonant preceding the
  vowel is oral or nasal. NA if these measurements are from a
  consonant.}
  \item{strF0}{Fundamental frequency of the voice, in Hz, as measured by
  STRAIGHT (Kawahara et al., 1998).}
  \item{sF0}{Fundamental frequency of the voice, in Hz, as measured by
  Snack.}
  \item{pF0}{Fundamental frequency of the voice, in Hz, as measured by
  Praat.}
  \item{shrF0}{Fundamental frequency of the voice, in Hz, as measured using
  the subharmonic-to-harmonic ratio method (Sun 2002).}
  \item{sF1}{Frequency of the first formant, in Hz, as measured by Snack.}
  \item{sB1}{Bandwidth of the first formant, in Hz, as measured by Snack.}
  \item{sF2}{Frequency of the second formant, in Hz, as measured by Snack.}
  \item{sB2}{Bandwidth of the second formant, in Hz, as measured by Snack.}
  \item{sF3}{Frequency of the third formant, in Hz, as measured by Snack.}
  \item{sB3}{Bandwidth of the third formant, in Hz, as measured by Snack.}
  \item{sF4}{Frequency of the fourth formant, in Hz, as measured by Snack.}
  \item{sB4}{Bandwidth of the fourth formant, in Hz, as measured by Snack.}
  \item{pF1}{Frequency of the first formant, in Hz, as measured by Praat.}
  \item{pF2}{Frequency of the second formant, in Hz, as measured by Praat.}
  \item{pF3}{Frequency of the third formant, in Hz, as measured by Praat.}
  \item{pF4}{Frequency of the fourth formant, in Hz, as measured by Praat.}
  \item{H1u}{Amplitude of the first harmonic of the voice.}
  \item{H1c}{Amplitude of the first harmonic of the voice. Corrected for the
  formant filter based on Iseli & Alwan (2004, 2006, 2007).}
  \item{H2u}{Amplitude of the second harmonic of the voice.}
  \item{H2c}{Amplitude of the second harmonic of the voice. Corrected for the
  formant filter.}
  \item{H4u}{Amplitude of the third harmonic of the voice.}
  \item{H4c}{Amplitude of the fourth harmonic of the voice. Corrected for the
  formant filter.}
  \item{A1u}{Amplitude of the harmonic closest to the first formant peak
  frequency, in decibels.}
  \item{A1c}{Amplitude of the harmonic closest to the first formant peak
  frequency, in decibels. Corrected for the formant filter.}
  \item{A2u}{Amplitude of the harmonic closest to the second formant peak
  frequency, in decibels.}
  \item{A2c}{Amplitude of the harmonic closest to the second formant peak
  frequency, in decibels. Corrected for the formant filter.}
  \item{A3u}{Amplitude of the harmonic closest to the third formant peak
  frequency, in decibels.}
  \item{A3c}{Amplitude of the harmonic closest to the third formant peak
  frequency, in decibels. Corrected for the formant filter.}
  \item{H1H2u}{Difference between H1 and H2, in decibels.}
  \item{H1H2c}{Difference between H1 and H2, in decibels. Corrected for the
  formant filter.}
  \item{H2H4u}{Difference between H2 and H4, in decibels.}
  \item{H2H4c}{Difference between H2 and H4, in decibels. Corrected for the
  formant filter.}
  \item{H1A1u}{Difference between H1 and A1, in decibels.}
  \item{H1A1c}{Difference between H1 and A1, in decibels. Corrected for the
  formant filter.}
  \item{H1A2u}{Difference between H1 and A2, in decibels.}
  \item{H1A2c}{Difference between H1 and A2, in decibels. Corrected for the
  formant filter.}
  \item{H1A3u}{Difference between H1 and A3, in decibels.}
  \item{H1A3c}{Difference between H1 and A3, in decibels. Corrected for the
  formant filter.}
  \item{Energy}{Root mean square energy.}
  \item{HNR05}{Harmonics-to-noise ratio between 0-500 Hz.}
  \item{HNR15}{Harmonics-to-noise ratio between 0-1500 Hz.}
  \item{HNR25}{Harmonics-to-noise ratio between 0-2500 Hz.}
  \item{HNR35}{Harmonics-to-noise ratio between 0-3500 Hz.}
  \item{CPP}{Cepstral peak prominence, in decibels.}
  \item{SHR}{Amplitude ratio between subharmonics and harmonics (Sun 2002).
  If there is a `0` in this column, please be aware that it sometimes
  indicates bad data.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
acoustics
}
\description{
Acoustic measurements from recordings of Bo, English, Gujarati, Hmong,
Luchun, Mandarin, Mazatec, Miao, Yi, and Zapotec. Measurements are taken at
nine timepoints within each phonetic segment (vowel or consonant), including
the pitch, formant filter, harmonic amplitudes, energy, and noise.
}
\details{
This documentation is based on the the spreadsheet key on the
\href{http://www.phonetics.ucla.edu/voiceproject/voice.html}{UCLA project website}.
Please see that page for more information about the corpus (including the
original audio recordings) and for citation and licensing instructions. For
more information on the acoustic measures, see the
\href{http://www.seas.ucla.edu/spapl/voicesauce/}{VoiceSauce documentation}.

This table is a version of the original acoustic master spreadsheet converted
to the
\href{https://r4ds.had.co.nz/tidy-data.html}{tidy data format} for use with
the tidyverse R packages. It has one row per timepoint, rather than one row
per filename, and one column per acoustic measure, rather than nine timepoint
columns and one overall mean for each measure. Each phonetic segment in the
data has a unique combination of \code{filename} and \code{segment_start}
values, and each measurement has a unique combination of \code{filename},
\code{segment_start}, and \code{timepoint} values.

The original spreadsheet had a column to indicate bad data in some cells. In
this version, all of these values have been replaced with NA, and the bad
data column has been removed. The authors also recommend further exclusions,
described on the second page of the original spreadsheet:

1. Exclude data where the STRAIGHT f0 measurement (\code{strF0}) differs from
 the Praat f0 measurement (\code{pF0}) by more than 200 Hz. See first example
 below.

2. Exclude all data from any file where the mean of the middle three Snack
F1 measurements (\code{sF1}) differs from the mean of the middle three Praat
F1 measurements (\code{pF1}) by more than 200 Hz. See second example below.
}
\examples{
data("acoustics", package = "voicequality")

# First suggested exclusion, when using measures that depend on pitch tracking
filter(acoustics, abs(strF0 - pF0) <= 200)

# Second suggested exclusion, when using measures that depend on formant estimates
filenames_to_include <- acoustics \%>\%
  filter(between(timepoint, 4, 6)) \%>\%
  group_by(filename, segment_start) \%>\%
  summarize(diff_F1 = abs(mean(sF1) - mean(pF1))) \%>\%
  filter(diff_F1 <= 200) \%>\%
  pull(filename)

filter(acoustics, filename \%in\% unique(filenames_to_include))

}
\keyword{datasets}
