% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bo_word_list}
\alias{bo_word_list}
\title{Word list for Bo}
\format{A data frame with 79 rows and 3 variables:
\describe{
  \item{transcription}{Phonetic transcription of the word. See PDF notes on
  the project website for IPA equivalents.}
  \item{english_gloss}{English gloss of the word.}
  \item{mandarin_gloss}{Mandarin gloss of the word.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
bo_word_list
}
\description{
List of words elicited in recordings of Bo for the UCLA voice quality
project.
}
\details{
Data collection credit: Jianjing Kuang of UCLA, fieldwork in two Bo villages
(Shizong and Xingfucun) in summer 2009.
}
\examples{
data("bo_word_list", package = "voicequality")

}
\keyword{datasets}
