% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egg}
\alias{egg}
\title{Electroglottography measurements for the UCLA voice quality project}
\format{A data frame with 133,254 rows and 32 variables: \describe{
  \item{speaker_id}{Speaker ID. Unique across all eight languages.}
  \item{speaker_number}{Speaker number or pseudonym. Unique within a language
  but not across languages.}
  \item{speaker_sex}{F for female or M for male.}
  \item{language}{Language that this word is from.}
  \item{language_variety}{Dialect of the language, or the location where this
  word was recorded.}
  \item{filename}{Filename of the corresponding EGG file. For a description
  of file naming conventions, see the PDF notes for each language on the
  project website.}
  \item{segment_start}{Start time of this phonetic segment in the EGG file,
  in milliseconds.}
  \item{segment_end}{End time of this phonetic segment in the EGG file, in
  milliseconds.}
  \item{duration}{Duration of this segment, in milliseconds.}
  \item{timepoint}{Time within the segment where these measurements were
  taken, as a number from 1 to 9. A 1 means that this row has measurements
  taken during the first ninth of the segment, a 2 means that it has
  measurements taken during the second ninth, etc.}
  \item{segment_type}{Whether the segment is a vowel or a consonant.}
  \item{vowel_quality}{Vowel quality for this segment. NA if these
  measurements are from a consonant.}
  \item{phonation}{Phonation type of the segment.}
  \item{vowel_nasality}{Whether the vowel is oral or nasal.}
  \item{consonant_aspiration}{Whether the consonant preceding the vowel is
  aspirated or unaspirated. If this segment is a consonant, then whether this
  segment is aspirated or unaspirated.}
  \item{tone_contour}{Sequence of tone targets for this tone type, using the
  standard IPA range with the numbers 1-5, where 1 is low, 3 is mid, and 5 is
  high. NA for languages without lexical tone.}
  \item{number_of_tone_targets}{Number of tone targets for this tone type,
  from 1 to 3.}
  \item{tone_category}{Label for tone contour, divided into four categories:
  high (55, 53, 45, 44, 535), mid (24, 33, 35, 31), low (22, 21, 11, 213,
  13), and big (51, 351, 15, 151, 153).}
  \item{language_specific_tone}{Tone label based on language-specific tone
  transcription conventions, which were also used to annotate tone in the
  original TextGrid file. For example, in Mandarin Chinese, the files were
  annotated with the traditional labels of Tone 1, 2, 3 or 4. The
  \code{tone_contour} and \code{tone_category} columns have the actual tone
  contours for each tone type.}
  \item{tone_misc}{...}
  \item{CQ}{Contact quotient, calculated with the standard method, which uses
  a pre-assigned percentage of the cycle height for the closure start time.}
  \item{CQ_H}{Contact quotient, calculated with a hybrid method, which uses
  the peak velocity time as the closure start time and a pre-assigned
  percentage of the cycle height for the closure end time.}
  \item{CQ_HT, CQ_PM}{Contact quotient, calculated using other methods.
  See the \href{http://www.appsobabble.com/functions/EGGWorks.aspx}{EGGWorks}
  website by Henry Tehrani for a description of these methods.}
  \item{min_Vel}{Cycle minimum velocity.}
  \item{min_Vel_Time}{Cycle minimum velocity time.}
  \item{peak_Vel}{Cycle peak velocity.}
  \item{peak_Vel_Time}{Cycle peak velocity time.}
  \item{SQ2_SQ1}{Duration of the closing slope, calculated as the time from
  when the cycle reaches 10\% of its minimum value to when the cycle reaches
  90\% of its minimum value.}
  \item{SQ4_SQ3}{Duration of the opening slope, calculated as the time from
  when the cycle reaches 90\% of its minimum value to when the cycle reaches
  10\% of its minimum value.}
  \item{ratio}{Ratio of closing slope to opening slope durations.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
  \url{http://phonetics.linguistics.ucla.edu/facilities/physiology/egg.htm}
  \url{http://www.appsobabble.com/functions/EGGWorks.aspx}
}
\usage{
egg
}
\description{
Electroglottography measurements from recordings of Bo, Gujarati, Hmong,
Luchun, Mandarin, Miao, Yi, and Zapotec. Measurements are taken at nine
timepoints within each phonetic segment (vowel or consonant).
}
\details{
This documentation is based on the the spreadsheet key on the
\href{http://www.phonetics.ucla.edu/voiceproject/voice.html}{UCLA project
website}. Please see that page for more information about the corpus
(including the original EGG recordings) and for citation and licensing
instructions. The description of the EGG measures is from the
\href{http://www.appsobabble.com/functions/EGGWorks.aspx}{EGGWorks}
documentation (requires registration).

This table is a version of the original EGG master spreadsheet converted to
the \href{https://r4ds.had.co.nz/tidy-data.html}{tidy data format} for use
with the tidyverse R packages. It has one row per timepoint, rather than one
row per filename, and one column per EGG measure, rather than nine timepoint
columns and one overall mean for each measure. Each phonetic segment in the
data has a unique combination of \code{filename} and \code{segment_start}
values, and each measurement has a unique combination of \code{filename},
\code{segment_start}, and \code{timepoint} values.

Be aware that the \code{min_Vel}, \code{peak_Vel}, \code{ratio},
\code{SQ2_SQ1}, and \code{SQ4_SQ3} columns contain extreme values that are
most likely bad data. Be sure to check the distribution of the data before
doing any analyses.
}
\examples{
data("egg", package = "voicequality")

}
\keyword{datasets}
