% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gujarati_word_list}
\alias{gujarati_word_list}
\title{Word list for Gujarati}
\format{A data frame with 75 rows and 5 variables:
\describe{
  \item{file_number}{The number used in the audio filenames to refer to this
  word. For example, if a filename is \code{2008-11-21-1330-01}, the last
  number (01) indicates that the audio file is a recording of the word in
  this table with a \code{file_number} of 01.}
  \item{target_pronunciation}{Expected pronunciation of the word, which is
  not necessarily the actual production of the speaker.}
  \item{dictionary_pronunciation}{Dictionary pronunciation of the word.}
  \item{target_vowel}{Transcription of the vowel (or vowel with preceding
  aspiration) that was extracted from this word.}
  \item{gloss}{English gloss of the word.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
gujarati_word_list
}
\description{
List of words elicited in recordings of Gujarati for the UCLA voice quality
project.
}
\details{
Data collection credit: Sameer Khan of UCLA, recordings all made in the
soundbooth of the UCLA Phonetics Lab, in 2008-2009.
}
\examples{
data("gujarati_word_list", package = "voicequality")

}
\keyword{datasets}
