% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmong_word_list}
\alias{hmong_word_list}
\title{Word list for Hmong}
\format{A data frame with 163 rows and 5 variables:
\describe{
  \item{index}{Index for the word, not used in the filenames.}
  \item{orthography}{Spelling of the word.}
  \item{IPA}{Phonetic transcription of the word in IPA.}
  \item{gloss}{English gloss of the word.}
  \item{language_variety}{Whether the word is from White Hmong or Green
  Hmong.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
hmong_word_list
}
\description{
List of words elicited in recordings of Hmong for the UCLA voice quality
project.
}
\details{
Data collection credit: Recordings were made by Christina Esposito of
Macalester College and Sherrie Yang of UCLA at the Hmong American Partnership
(St. Paul, Minnesota) and the Immmanuel Hmong Lutheran Church (St. Paul,
Minnesota) in summer 2008.
}
\examples{
data("hmong_word_list", package = "voicequality")

}
\keyword{datasets}
