% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luchun_word_list}
\alias{luchun_word_list}
\title{Word list for Luchun}
\format{A data frame with 108 rows and 2 variables:
\describe{
  \item{index}{Index for the word, not used in the filenames.}
  \item{transcription}{Phonetic transcription of the word. See PDF notes on
  the project website for IPA equivalents.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
luchun_word_list
}
\description{
List of words elicited in recordings of Luchun by Professor Jiangping Kong of
Peking University during fieldwork in Luchun village in summer 2009. Included
in the UCLA voice quality project.
}
\examples{
data("luchun_word_list", package = "voicequality")

}
\keyword{datasets}
