% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mazatec_word_list}
\alias{mazatec_word_list}
\title{Word list for Mazatec}
\format{A data frame with 82 rows and 6 variables:
\describe{
  \item{filename}{Transcription used in the audio filenames to refer to this
  word.}
  \item{IPA}{Phonetic transcription of the word in IPA.}
  \item{gloss}{English gloss of the word.}
  \item{orthography}{Spelling of the word. Not available for 1984
  recordings.}
  \item{index}{Index number from the UCLA Phonetic Database.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
mazatec_word_list
}
\description{
List of words elicited in recordings of Mazatec by Paul Kirk in 1982 and
brought to UCLA in 1984; and by Paul Kirk, Peter Ladefoged, and others from
UCLA in 1993. Included in the UCLA voice quality project.
}
\examples{
data("mazatec_word_list", package = "voicequality")

}
\keyword{datasets}
