% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miao_filename_list}
\alias{miao_filename_list}
\title{Miao filename metadata}
\format{A data frame with 1,697 rows and 6 variables:
\describe{
  \item{speaker}{Speaker name.}
  \item{code}{Code for speaker.}
  \item{transcription}{Phonetic transcription of the word. See PDF notes on
  the project website for IPA equivalents.}
  \item{gloss}{English gloss of the word.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
miao_filename_list
}
\description{
Metadata for audio files of Miao recordings in the UCLA voice quality
project.
}
\details{
Data collection credit: Jianjing Kuang of UCLA, in Shidong Kou
(Shih-Tung-K'ou), Taijiang (T'ai-Kung) county of Guizhou (Kweichow) province,
China, in summer 2011.
}
\examples{
data("miao_filename_list", package = "voicequality")

}
\seealso{
Other Miao word lists: \code{\link{miao_minimal_sets}},
  \code{\link{miao_other_words}}
}
\concept{Miao word lists}
\keyword{datasets}
