% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miao_minimal_sets}
\alias{miao_minimal_sets}
\title{Word list for Miao minimal sets}
\format{A data frame with 87 rows and 4 variables:
\describe{
  \item{tone}{Sequence of tone targets for this tone type, using the
  standard IPA range with the numbers 1-5, where 1 is low, 3 is mid, and 5 is
  high.}
  \item{IPA}{Phonetic transcription of the word in IPA.}
  \item{gloss}{English gloss of the word.}
  \item{minimal_set_id}{Identifier for all members of a minimal set. For
  example, all rows with \code{minimal_set_id} of 1 belong to a minimal set,
  all rows with \code{minimal_set_id} of 2 belong to a different minimal set,
  etc.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
miao_minimal_sets
}
\description{
List of minimal sets elicited in recordings of Miao for the UCLA voice
quality project.
}
\details{
Data collection credit: Jianjing Kuang of UCLA, in Shidong Kou
(Shih-Tung-K'ou), Taijiang (T'ai-Kung) county of Guizhou (Kweichow) province,
China, in summer 2011.
}
\examples{
data("miao_minimal_sets", package = "voicequality")

}
\seealso{
Other Miao word lists: \code{\link{miao_filename_list}},
  \code{\link{miao_other_words}}
}
\concept{Miao word lists}
\keyword{datasets}
