% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miao_other_words}
\alias{miao_other_words}
\title{Word list for other Miao words}
\format{A data frame with 39 rows and 3 variables:
\describe{
  \item{IPA}{Phonetic transcription of the word in IPA.}
  \item{tone}{Sequence of tone targets for this tone type, using the
  standard IPA range with the numbers 1-5, where 1 is low, 3 is mid, and 5 is
  high.}
  \item{gloss}{English gloss of the word.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
miao_other_words
}
\description{
List of other words, besides minimal sets, elicited in recordings of Miao for
the UCLA voice quality project.
}
\details{
Data collection credit: Jianjing Kuang of UCLA, in Shidong Kou
(Shih-Tung-K'ou), Taijiang (T'ai-Kung) county of Guizhou (Kweichow) province,
China, in summer 2011.
}
\examples{
data("miao_other_words", package = "voicequality")

}
\seealso{
Other Miao word lists: \code{\link{miao_filename_list}},
  \code{\link{miao_minimal_sets}}
}
\concept{Miao word lists}
\keyword{datasets}
