% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yi_tense_lax}
\alias{yi_tense_lax}
\title{Word list for Yi tense/lax minimal pairs}
\format{A data frame with 82 rows and 5 variables:
\describe{
  \item{vowel}{Transcription of the vowel that was extracted from this word.}
  \item{tense_lax}{Whether the vowel is tense or lax.}
  \item{minimal_pair_id}{Identifier for all members of a minimal set. For
  example, the rows with \code{minimal_pair_id} of 1 belong to a minimal
  pair, the rows with \code{minimal_pair_id} of 2 belong to a different
  minimal pair, etc.}
  \item{IPA}{Phonetic transcription of the word in IPA.}
  \item{english_gloss}{English gloss of the word.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
yi_tense_lax
}
\description{
List of tense/lax minimal pairs elicited in recordings of Yi for the UCLA
voice quality project.
}
\details{
Data collection credit: Jianjing Kuang of UCLA, fieldwork in two southern Yi
villages (Xinping and Jiangcheng) in summer 2009.
}
\examples{
data("yi_tense_lax", package = "voicequality")

}
\seealso{
Other Yi word lists: \code{\link{yi_tones}},
  \code{\link{yi_vowels}}
}
\concept{Yi word lists}
\keyword{datasets}
