% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yi_tones}
\alias{yi_tones}
\title{Word list for Yi showing the phonation contrast on different tones}
\format{A data frame with 86 rows and 4 variables:
\describe{
  \item{tone}{Sequence of tone targets for this tone type, using the
  standard IPA range with the numbers 1-5, where 1 is low, 3 is mid, and 5 is
  high.}
  \item{tense_lax}{Whether the vowel is tense or lax.}
  \item{minimal_set_id}{Identifier for all members of a minimal set. For
  example, all rows with \code{minimal_set_id} of 1 belong to a minimal
  set, all rows with \code{minimal_set_id} of 2 belong to a different
  minimal set, etc.}
  \item{english_gloss}{English gloss of the word.}
  \item{IPA}{Phonetic transcription of the word in IPA.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
yi_tones
}
\description{
List of tense/lax minimal pairs elicited in recordings of Yi for the UCLA
voice quality project.
}
\details{
Data collection credit: Jianjing Kuang of UCLA, fieldwork in two southern Yi
villages (Xinping and Jiangcheng) in summer 2009.
}
\examples{
data("yi_tones", package = "voicequality")

}
\seealso{
Other Yi word lists: \code{\link{yi_tense_lax}},
  \code{\link{yi_vowels}}
}
\concept{Yi word lists}
\keyword{datasets}
