% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yi_vowels}
\alias{yi_vowels}
\title{Word list for Yi showing the phonation contrast on different vowels}
\format{A data frame with 38 rows and 4 variables:
\describe{
  \item{vowel}{Transcription of the vowel that was extracted from this word.}
  \item{minimal_pair_id}{Identifier for all members of a minimal pair. For
  example, the rows with \code{minimal_pair_id} of 1 belong to a minimal
  pair, the rows with \code{minimal_pair_id} of 2 belong to a different
  minimal pair, etc.}
  \item{english_gloss}{English gloss of the word.}
  \item{IPA}{Phonetic transcription of the word in IPA.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
yi_vowels
}
\description{
List of tense/lax minimal pairs elicited in recordings of Yi for the UCLA
voice quality project.
}
\details{
Data collection credit: Jianjing Kuang of UCLA, fieldwork in two southern Yi
villages (Xinping and Jiangcheng) in summer 2009.
}
\examples{
data("yi_vowels", package = "voicequality")

}
\seealso{
Other Yi word lists: \code{\link{yi_tense_lax}},
  \code{\link{yi_tones}}
}
\concept{Yi word lists}
\keyword{datasets}
