% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zapotec_word_list}
\alias{zapotec_word_list}
\title{Word list for Zapotec}
\format{A data frame with 48 rows and 5 variables:
\describe{
  \item{index}{Index used in the audio filenames to refer to this word. For
  example, if a filename is \code{speaker3M_aC34}, the last number (34)
  indicates that the audio file is a recording of the word in this table with
  an \code{index} of 34.}
  \item{IPA}{Phonetic transcription of the word in IPA.}
  \item{orthography}{Spelling of the word.}
  \item{english_gloss}{English gloss of the word.}
  \item{spanish_gloss}{Spanish gloss of the word.}
  \item{phonation}{Phonation type of the segment.}
}}
\source{
\url{http://www.phonetics.ucla.edu/voiceproject/voice.html}
}
\usage{
zapotec_word_list
}
\description{
List of words elicited in recordings of Zapotec for the UCLA voice quality
project.
}
\details{
Data collection credit: Christina Esposito, with assistance from Alvaro Luna,
at various locations in the Koreatown area of Los Angeles, in summer 2010.
}
\examples{
data("zapotec_word_list", package = "voicequality")

}
\keyword{datasets}
